+++ Core/Random/Mt19937.cpp
@@ -5,7 +5,7 @@
 namespace relleums {
 namespace Random {
 
-Mt19937::Mt19937(const uint32_t seed) {
+Mt19937::Mt19937(const uint64_t seed) {
     init_inverse_maximum();
     set_seed(seed);
 }
@@ -19,7 +19,7 @@ double Mt19937::uniform() {
     return static_cast<double>(pRNG_mt19937())*inv_max;
 }
 
-void Mt19937::set_seed(const uint32_t seed) {
+void Mt19937::set_seed(const uint64_t seed) {
     this->seed_ = seed;
     pRNG_mt19937.seed(this->seed_);
 }
